+++ client/src/components/Workflow/List/WorkflowIndicators.vue
